reename PriceOracleMoCTest to PriceOracleProxyTest
